chore: adapt consolidation-flow to new way of doing it on AC
To do consolidation, we now call an endpoint on the minigate instead of
submitting a tx to AC. The flow in the consolidations-module has thus
been adjusted accordingly. The results are then processed based on events we
receive (sign and l1-confirm), and the job-processors have also been
adapted to hadle the differences from normal events.